requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==2.3.3
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pychrome==0.2.4
PyYAML==6.0.1
requests==2.32.3
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==2.3.7
